backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

langchain==0.3.18
langchain-groq==0.2.2
langchain-huggingface==0.1.2
chromadb>=0.5.0
sentence-transformers>=2.6.0
groq

pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
wikipedia==1.4.0

pydantic>=2.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
